@@ -1,13 +1,12 @@
-import Vue, { VNode } from "vue";
-
-declare global {
-  namespace JSX {
-    // tslint:disable no-empty-interface
-    interface Element extends VNode {}
-    // tslint:disable no-empty-interface
-    interface ElementClass extends Vue {}
-    interface IntrinsicElements {
-      [elem: string]: any;
-    }
-  }
-}
+// import { VNode, ComponentInternalInstance } from "vue";
+// declare global {
+// namespace JSX {
+// tslint:disable no-empty-interface
+// interface Element extends VNode {}
+// tslint:disable no-empty-interface
+// interface ElementClass extends ComponentInternalInstance {}
+// interface IntrinsicElements {
+// [elem: string]: any;
+// }
+// }
+// }
